Apps script clean up
Refactored stuff to make it neater and understandable
Made separate firebase functions files for Acuity and Bookings for easier readability.
Change app name and icon in index.html and manifest
