+++ apps-script/appsscript.json
@@ -3,6 +3,7 @@
   "dependencies": {
   },
   "exceptionLogging": "STACKDRIVER",
+  "runtimeVersion": "V8",
   "executionApi": {
     "access": "MYSELF"
   }
