+++ firebase/public/manifest.json
@@ -1,11 +1,11 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Fizz Kidz Bookings",
+  "name": "Fizz Kidz Booking System",
   "icons": [
     {
-      "src": "favicon.ico",
+      "src": "fizz_logo.png",
       "sizes": "64x64 32x32 24x24 16x16",
-      "type": "image/x-icon"
+      "type": "image/png"
     }
   ],
   "start_url": ".",
